Refactor mobile navigation components to use MobileBottomBar
- Replaced the existing MobileMoreMenu with a new MobileBottomBar component across AskPage, BestPage, FrontPage, JobsPage, ShowPage, and HNLiveTerminal for improved code reusability and maintainability.
- Consolidated mobile navigation buttons, including Home, Front Page, Search, and Settings, into the new MobileBottomBar.
- Removed redundant state management and click outside functionality from individual pages, centralizing it within the MobileBottomBar component.
- Enhanced user experience by streamlining mobile navigation and ensuring consistent layout and functionality across all pages.
